updated the readme to reflect my ownership.
Mainly this is just me bogarting it to stop others from bogarting it or stealing it from me. I plan to release some final product, derived from this, under: a more permissive license, at a later date. But until then this is mine *evil laugh*.

I'm also a bit paranoid when it comes to my art, which this, I consider to be art. so yes.

I will be getting help from people on Fiverr but they are technically my employees so their contributions to this project are still ultimately mine, and belong to me as goods I paid for.