Added the Empathy Engine
The empathy engine sorts through how it wants to make the user feel, and then selects the best sentiment to send to an as of yet nonexistent generative model which will be rewarded for correctly guessing the sentiment and punished for incorrectly guessing the sentiment, proportional to the distance from the target sentiment, based on user's feedback. This will teach the robot how to properly make the user happy. It all starts with the empathy engine, that outputs the target emotion. If the user provides feedback that the input did reach the target emotion the chatbot is rewarded.  the distance isn't the only metric used to determine how to reward and punish the chatbot too, the discomfort level is a punishment multiplier. If the user's feedback uncovers a new emotion, the user will be prompted to name that emotion and give examples of it made them feel and it will train the sentiment analysis AI on it, as well as the Text Generation AI on it. They'll both have to be siamese neural networks capable of one-shot learning.