I changed the emodex set_heart func to a float
I did this because the user's heart is in integers but the robot's heart should be in float values, as any perceived emotional intent from the texts it receives effects it in a 2:5 ration average 2 of the emotional intent, 3 of the current emotion, added together, divided by 5. The robot starts out as the user's initial emotions, then. as each of them talk the emotional intent the robot expresses and the emotional intent the human expresses change the robot's emotional state.