Poly Vagal Impulse Module
Completely reworked the empathy engine, split the emodex into a user emodex, and a bot emodex, then for bot on bot communications, we can have them have empathy for each other, or have empathy for the user. Oh and I spent Faaaaaaaaaaaaaaaaaaaaaaaaaaaaaar too long on the Poly Vagal Impulse Module, which is just a fancy way of saying a calculated delayer for unprompted messages based on the bot's sum emotional state. But getting it to update immediately when the emotions change, was the tricky part.
Still not sure if I've got it fine tuned the way I want it but it'll do for now.
Allowing the calculated delay to update immediately as the bot's emotional state changes, allows it to respond right away, and then respond to its own response, when it is ready.

At some point I wanna include an inner monologue module that goes something along the lines of "you are feeling [emotion], what do you think about that?" and prompt chain it with the chat history in some way.

and then when it is time to respond to a user: "you are thinking about [response to above prompt] and name:[name], said: [text message], please respond".